refactor(default_values_dictionaries): Rename file to "constants" as more reflective of file contents and consistent with use in the package.
