+++ src/nasem_dairy/__init__.py
@@ -5,7 +5,7 @@ __version__ = version("nasem_dairy")
 from nasem_dairy.ration_balancer.ration_balancer_functions import get_feed_rows_feedlibrary, check_coeffs_in_coeff_dict, read_csv_input, read_infusion_input
 from nasem_dairy.ration_balancer.ModelOutput import ModelOutput
 from nasem_dairy.ration_balancer.execute_model import execute_model
-from nasem_dairy.ration_balancer.default_values_dictionaries import coeff_dict, infusion_dict, MP_NP_efficiency_dict, mPrt_coeff_list
+from nasem_dairy.ration_balancer.constants import coeff_dict, infusion_dict, MP_NP_efficiency_dict, mPrt_coeff_list
 from nasem_dairy.NASEM_equations.DMI_equations import (
     calculate_Kb_LateGest_DMIn,
     calculate_An_PrePartWklim,
